requirements.txt
vanna[sqlite]>=2.0.0
vanna[gemini, sqlite]
fastapi
uvicorn[standard]
plotly
pandas
python-dotenv
google-genai
faker